backend/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain-google-genai
langchain-community
duckduckgo-search
beautifulsoup4 